refactor minor heaps allocation
This commit refactor the minor heaps allocation process by first rewriting the logic to not
rely on bitmasks and instead move it to a model closer to trunk, introducing a Minor_heap_max configuration variable.
Some bits are removed: the arm64 backend used to expose Minor_heap_align_bits in config.h to the emitter, which is removed for now.
The Is_minor/Is_young changes are now unified as only Is_young (is the value in *any* minor heap at all?) matters.
In a follow up commit, Is_minor will be renamed definitely to Is_yound to follow suit.